+++ test/dummy/config/initializers/nuclear_secrets.rb
@@ -1,7 +1,6 @@
 NuclearSecrets.configure do |config|
   config.required_secrets = {
-    #secret_key_base: String,
+    secret_key_base: String,
     secret_token: NilClass,
-    secret_key_base: Proc.new { false } 
   }
 end
